+++ feature-organizer/src/commonMain/kotlin/org/pointyware/timer/organizer/ui/OrganizerElement.kt
@@ -1,4 +1,4 @@
-package org.pointyware.timer.organizer
+package org.pointyware.timer.organizer.ui
 
 /**
  * Used for presentation of an organizational hierarchy.
